Made the file more organized
